backend/requirements.txt
apache-airflow
psycopg2-binary
python-dotenv
pandas
prophet
requests
matplotlib